dags/requirements.txt
alembic==1.6.5
anyio==3.2.1
apache-airflow==2.1.1
apache-airflow-providers-docker==2.0.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
blinker==1.4
cached-property==1.5.2
cattrs==1.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
contextvars==2.4
croniter==1.0.15
cryptography==3.4.7
dataclasses==0.8
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docker==5.0.0
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
immutables==0.15
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
numpy==1.19.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pandas==1.1.5
pendulum==2.1.2
pep562==1.1
prison==0.1.3
psutil==5.8.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
rich==10.5.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
typing==3.7.4.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.5.0
